All tiles of the first bank can now be used

+++ src/main/java/io/github/mibi88/mibinestools/NametableEditor.java
@@ -85,7 +85,7 @@ public class NametableEditor extends Editor {
         nametableViewer.setCHR(chrData);
     }
     
-    public void setCurrentTile(byte currentTile) {
+    public void setCurrentTile(int currentTile) {
         nametableViewer.setCurrentTile(currentTile);
     }
